refactor: use a layout for the main app
