+++ packages/cysync-core/src/pages/MainApp/Components/index.ts
@@ -1,2 +1,2 @@
 export * from './AssetAllocation';
-export * from './SideBar';
+export * from './MainAppLayout';
