+++ packages/cysync-core/src/pages/MainApp/Portfolio.tsx
@@ -1,13 +1,9 @@
-import { Container, Flex } from '@cypherock/cysync-ui';
 import React, { FC, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 
 import { openWalletActionsDialog } from '~/actions';
-import { Topbar } from '~/components';
-import { AssetAllocation } from '~/pages/MainApp/Components/AssetAllocation';
+import { AssetAllocation, MainAppLayout } from '~/pages/MainApp/Components';
 import { selectLanguage, useAppSelector } from '~/store';
-
-import { SideBar } from './Components';
 
 export const Portfolio: FC = () => {
   const { strings } = useAppSelector(selectLanguage);
@@ -18,12 +14,8 @@ export const Portfolio: FC = () => {
   }, []);
 
   return (
-    <Container height="screen" display="flex">
-      <SideBar />
-      <Flex direction="column" grow={1} $alignSelf="start">
-        <Topbar title={strings.portfolio.title} />
-        <AssetAllocation />
-      </Flex>
-    </Container>
+    <MainAppLayout title={strings.portfolio.title}>
+      <AssetAllocation />
+    </MainAppLayout>
   );
 };
