+++ packages/cysync-core/src/pages/MainApp/Wallet.tsx
@@ -1,11 +1,9 @@
-import { Container, Flex } from '@cypherock/cysync-ui';
 import React, { FC } from 'react';
 
-import { Topbar } from '~/components';
 import { useQuery } from '~/hooks';
 import { selectLanguage, selectWallets, useAppSelector } from '~/store';
 
-import { SideBar } from './Components';
+import { MainAppLayout } from './Components';
 
 export const Wallet: FC = () => {
   const { strings } = useAppSelector(selectLanguage);
@@ -15,12 +13,5 @@ export const Wallet: FC = () => {
     wallet => wallet.__id === query.get('id'),
   )?.name;
 
-  return (
-    <Container height="screen" display="flex">
-      <SideBar />
-      <Flex direction="column" grow={1} $alignSelf="start">
-        <Topbar title={`${strings.wallet.title} ${walletName}`} />
-      </Flex>
-    </Container>
-  );
+  return <MainAppLayout title={`${strings.wallet.title} ${walletName}`} />;
 };
